Move delete views to app.main.views.create_a_requirement.delete
