+++ app/main/views/create_a_requirement/__init__.py
@@ -1,2 +1,3 @@
 from . import create
+from . import delete
 from . import publish
